+++ internal/runtime/internal/controller/node_config_foreach.go
@@ -16,7 +16,7 @@ import (
 	"github.com/prometheus/client_golang/prometheus"
 
 	"github.com/grafana/alloy/internal/component"
-	"github.com/grafana/alloy/internal/dynamic/foreach"
+	"github.com/grafana/alloy/internal/foreach"
 	"github.com/grafana/alloy/internal/runner"
 	"github.com/grafana/alloy/internal/runtime/logging/level"
 	"github.com/grafana/alloy/syntax/ast"
