+++ internal/runtime/source.go
@@ -6,7 +6,7 @@ import (
 	"sort"
 	"strings"
 
-	"github.com/grafana/alloy/internal/dynamic/foreach"
+	"github.com/grafana/alloy/internal/foreach"
 	"github.com/grafana/alloy/internal/static/config/encoder"
 	"github.com/grafana/alloy/syntax/ast"
 	"github.com/grafana/alloy/syntax/diag"
